moved a bunch of code around
