+++ signals/generators/ios/objc_generator/__init__.py
@@ -1,0 +1,1 @@
+__author__ = 'Alex'
